tag increment

+++ Sources/PrepDataTypes/Server/Components/FoodType.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-public enum FoodType: Int, Codable {
+public enum FoodType: Int16, Codable {
     case rawFood = 1
     case plate
     case recipe
